MUSARK-108 Moving domain classes to common
