+++ common/src/main/scala/no/uio/musit/microservices/common/domain/MusitSearch.scala
@@ -16,7 +16,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package no.uio.musit.microservices.time.domain
+package no.uio.musit.microservices.common.domain
 
 case class MusitSearch(searchMap: Map[String, Option[String]], searchStrings: List[String])
