+++ service_time/app/no/uio/musit/microservices/time/service/TimeService.scala
@@ -18,6 +18,7 @@
  */
 package no.uio.musit.microservices.time.service
 
+import no.uio.musit.microservices.common.domain.{MusitError, MusitFilter}
 import no.uio.musit.microservices.time.domain._
 import org.joda.time.DateTime
 import org.joda.time.DateTime.now
@@ -26,16 +27,12 @@ trait TimeService {
   
   def convertToNow(maybeFilter: Option[MusitFilter]): Either[MusitError, MusitTime] =
     maybeFilter.map(f => resolveNow(f)).getOrElse(Right(fromDateTime(now)))
-    
-  def resolveNow(filter: MusitFilter) = filter match {
-    case MusitTimeFilter =>
-      Right(MusitTime(time = Some(now.toLocalTime)))
-    case MusitDateFilter =>
-      Right(MusitTime(date = Some(now.toLocalDate)))
-    case MusitDateTimeFilter =>
-      Right(fromDateTime(now))
-    case MusitFilter(other) =>
-      Left(MusitError(message = "Only supports empty filter or filter on time, date or time and date"))
+
+  def resolveNow(filter: MusitFilter): Either[MusitError, MusitTime] = filter match {
+    case MusitTimeFilter => Right(MusitTime(time = Some(now.toLocalTime)))
+    case MusitDateFilter => Right(MusitTime(date = Some(now.toLocalDate)))
+    case MusitDateTimeFilter => Right(fromDateTime(now))
+    case MusitFilter(other) => Left(MusitError(message = "Only supports empty filter or filter on time, date or time and date"))
   }
   
   def fromDateTime(dateTime: DateTime): MusitTime =
