+++ service_time/conf/routes
@@ -3,4 +3,4 @@
 # ~~~~
 
 # Setup endpoints
-GET     /v1/now                    no.uio.musit.microservices.time.resource.TimeResource.now(filter: Option[no.uio.musit.microservices.time.domain.MusitFilter], search: Option[no.uio.musit.microservices.time.domain.MusitSearch])
+GET     /v1/now                    no.uio.musit.microservices.time.resource.TimeResource.now(filter: Option[no.uio.musit.microservices.common.domain.MusitFilter], search: Option[no.uio.musit.microservices.common.domain.MusitSearch])
