+++ service_time/src/it/scala/TimeIntegrationTest.scala
@@ -17,13 +17,13 @@
  *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 import no.uio.musit.microservices.common.PlayTestDefaults
+import no.uio.musit.microservices.common.domain.MusitError
 import org.scalatest.concurrent.ScalaFutures
 import play.api.test.Helpers._
 import org.scalatestplus.play.{OneServerPerSuite, PlaySpec}
 import play.api.libs.json.Json
 import play.api.libs.ws.WS
 import no.uio.musit.microservices.time.domain.MusitTime
-import no.uio.musit.microservices.time.domain.MusitError
 
 /**
  * add your integration spec here.
